src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
docopt==0.6.2
idna==3.3
lxml==4.6.4
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
pipreqs==0.4.11
pycodestyle==2.7.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
yarg==0.1.9
